# api/openapi.json
{
  "components": {
    "schemas": {
      "AddressSummary": {
        "additionalProperties": false,
        "properties": {
          "address": {
            "description": "BLE MAC address as observed.",
            "title": "Address",
            "type": "string"
          },
          "first_observed_utc": {
            "format": "date-time",
            "title": "First Observed Utc",
            "type": "string"
          },
          "last_observed_utc": {
            "format": "date-time",
            "title": "Last Observed Utc",
            "type": "string"
          },
          "local_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Local Name"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "observations": {
            "title": "Observations",
            "type": "integer"
          },
          "scanners": {
            "items": {
              "type": "string"
            },
            "title": "Scanners",
            "type": "array"
          },
          "services": {
            "items": {
              "type": "string"
            },
            "title": "Services",
            "type": "array"
          },
          "vendors": {
            "items": {
              "type": "string"
            },
            "title": "Vendors",
            "type": "array"
          }
        },
        "required": [
          "address",
          "name",
          "local_name",
          "vendors",
          "services",
          "scanners",
          "observations",
          "first_observed_utc",
          "last_observed_utc"
        ],
        "title": "AddressSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_AddressSummary_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AddressSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor; pass to ?cursor= for the next page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[AddressSummary]",
        "type": "object"
      },
      "Page_Observation_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/Observation"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Opaque cursor; pass to ?cursor= for the next page.",
            "title": "Next Cursor"
          }
        },
        "required": [
          "items"
        ],
        "title": "Page[Observation]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Read-only HTTP API over the consolidated home-security archive.",
    "title": "Home Security API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/addresses": {
      "get": {
        "operationId": "list_addresses_addresses_get",
        "parameters": [
          {
            "description": "Substring match on address, name, or local_name.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on address, name, or local_name.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "scanner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scanner Id"
            }
          },
          {
            "description": "Only addresses last seen at or after this UTC instant.",
            "in": "query",
            "name": "seen_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Only addresses last seen at or after this UTC instant.",
              "title": "Seen Since"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "last_seen",
              "enum": [
                "last_seen",
                "first_seen",
                "observations"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AddressSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Paginated, filterable list of addresses",
        "tags": [
          "addresses"
        ]
      }
    },
    "/addresses/{address}": {
      "get": {
        "operationId": "address_detail_addresses__address__get",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "examples": [
                "AA:BB:CC:DD:EE:FF"
              ],
              "maxLength": 64,
              "minLength": 1,
              "title": "Address",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AddressSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Address not present in archive"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Single-address summary",
        "tags": [
          "addresses"
        ]
      }
    },
    "/addresses/{address}/observations": {
      "get": {
        "operationId": "address_observations_addresses__address__observations_get",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Address",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Observation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Observation timeline for one address",
        "tags": [
          "addresses"
        ]
      }
    },
...
    "/observations": {
      "get": {
        "operationId": "list_observations_observations_get",
        "parameters": [
          {
            "in": "query",
            "name": "scanner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scanner Id"
            }
          },
          {
            "in": "query",
            "name": "address",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Address"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_Observation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent observations across all scanners",
        "tags": [
          "observations"
        ]
      }
    },
...
    "/wifi/addresses": {
      "get": {
        "operationId": "list_addresses_wifi_addresses_get",
        "parameters": [
          {
            "description": "Substring match on MAC.",
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Substring match on MAC.",
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "scanner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scanner Id"
            }
          },
          {
            "description": "Filter to randomized (true) or stable (false) MACs.",
            "in": "query",
            "name": "randomized",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter to randomized (true) or stable (false) MACs.",
              "title": "Randomized"
            }
          },
          {
            "in": "query",
            "name": "seen_since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Seen Since"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "last_seen",
              "enum": [
                "last_seen",
                "first_seen",
                "observations"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WifiAddressSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Paginated, filterable list of Wi-Fi transmitter MACs",
        "tags": [
          "wifi"
        ]
      }
    },
    "/wifi/addresses/{address}": {
      "get": {
        "operationId": "address_detail_wifi_addresses__address__get",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "examples": [
                "a4:83:e7:00:11:22"
              ],
              "maxLength": 64,
              "minLength": 1,
              "title": "Address",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WifiAddressSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Address not present in archive"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Single Wi-Fi MAC summary",
        "tags": [
          "wifi"
        ]
      }
    },
    "/wifi/addresses/{address}/observations": {
      "get": {
        "operationId": "address_observations_wifi_addresses__address__observations_get",
        "parameters": [
          {
            "in": "path",
            "name": "address",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 1,
              "title": "Address",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "since",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Since"
            }
          },
          {
            "in": "query",
            "name": "until",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Until"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_WifiObservation_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Frame timeline for one Wi-Fi MAC",
        "tags": [
          "wifi"
        ]
      }
    },
...
  }
}